Updated python script and Jupyter notebook on chi-matrix
